4 lab/app/requirements.txt
pandas==1.3.4
psycopg2-binary==2.9.3
flask==2.2.4
flask_sqlalchemy==3.0.3
flask_migrate==4.0.4
redis